Archimedes2.0/requirements.txt
python-dotenv==1.1.0
duckdb==1.1.3
firebase-admin==6.7.0
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
icecream==2.1.4
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.12.3
jsonpointer==2.1
keras==3.5.0
mplfinance==0.12.10b0
mysql-connector==2.2.9
mysql-connector-python==9.2.0
numpy==1.26.4
oracledb==3.0.0
pandas==2.2.3
pandas-datareader==0.10.0
pandas-stubs==2.2.3.241126
pdfminer==20191125
pep8==1.7.1
pipreq==0.4
pipreqs==0.5.0
ply==3.11
polars==1.20.0
psycopg2==2.9.10
pydantic==2.11.0
pymongo==4.8.0
pymysql==1.1.1
pytest==8.3.5
pytest-snapshot==0.9.0
scipy==1.15.2
seaborn==0.12.2
scramp==1.4.5
sqlparse==0.5.3
virtualenv==20.26.5
XlsxWriter==3.2.2
